Clean up template engine after extension modularization
